First upload
It is the first script, it is not generic but a simple system made with 2 underground cables is used as example.
EMTP data file is generated from a FEMM project using the Matlab API of FEMM software. This data file can then be imported inside the cable data block of EMTP software.